finalizing our API

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -29,17 +29,21 @@ class TicketController extends ApiController
 
     public function store(StoreTicketRequest $request)
     {
-        try {
 
-            $this->isAble('store', Ticket::class);
+        if($this->isAble('store', Ticket::class))
+        {
 
             return new TicketResource(Ticket::create($request->mappedAttribute()));
 
-        }catch (AuthorizationException $ex){
-
-            return $this->error('You are not authorized to update this resource', 401);
-
         }
+
+
+
+
+
+        return $this->error('You are not authorized to update this resource', 401);
+
+
     }
 
 
@@ -75,19 +79,18 @@ class TicketController extends ApiController
 
             //policy
 
-            $this->isAble('update', $ticket);
+            if($this->isAble('update', $ticket))
+            {
+                $ticket->update($request->mappedAttribute());
 
-            $ticket->update($request->mappedAttribute());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to update this resource', 401);
 
         } catch (ModelNotFoundException){
 
             return $this->error('Ticket cannot be found.', 404);
-
-        }catch (AuthorizationException $ex){
-
-            return $this->error('You are not authorized to update this resource', 401);
 
         }
 
@@ -102,11 +105,14 @@ class TicketController extends ApiController
 
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('replace', $ticket);
+            if($this->isAble('replace', $ticket))
+            {
+                $ticket->update($request->mappedAttribute());
 
-            $ticket->update($request->mappedAttribute());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
+            return $this->error('You are not authorized to update this resource', 401);
 
         } catch (ModelNotFoundException){
 
@@ -122,11 +128,17 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            $this->isAble('delete', $ticket);
+            if($this->isAble('delete', $ticket))
+            {
 
-            $ticket->delete();
+                $ticket->delete();
 
-            return $this->ok('Ticket deleted successfully');
+                return $this->ok('Ticket deleted successfully');
+            }
+
+
+            return $this->error('You are not authorized to delete that resource', 401);
+
 
         } catch (ModelNotFoundException $exception){
